% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateGenotypes.R
\name{translateGenotypes}
\alias{translateGenotypes}
\title{Translate genotypes}
\usage{
translateGenotypes(input, ref_tbl, long = FALSE, output = NA, ...)
}
\arguments{
\item{input}{A data.frame or a (relative) path to a file of alleles to be
translated. Fixed columns are
name of the laboratory, which should match that from \code{ref_tbl}
and sample name (this is for your viewing pleasure only). The varying
columns represent genotype. Make sure columns match names from the
\code{ref_tbl}.}

\item{ref_tbl}{A data.frame which holds the translation table.
Structure is fixed and the columns are \code{lab_from}, \code{locus},
\code{allele_from}, \code{allele_ref} and \code{delta}. First two columns
are self explanatory. Columns that start with \code{allele_} have actual
allele values in laboratory of question and the reference (right now the
reference is laboratory from Slovenia). The last column is the amount
an allele should be shifted relative to the reference.}

\item{long}{Logical. If \code{TRUE}, the result will be returned as a long
table instead of wide. If \code{FALSE} (default), wide table will be
provided.}

\item{output}{Character. Relative or absolute path to the file the data
should be written to. The result will have tab separated columns and no row
names. Default is \code{NA}.}

\item{...}{Parameters passed to \link[readxl]{read_excel}.}
}
\description{
Translate genotypes
}
\details{
The shift (\code{delta}) can be possibly calculated for some loci.
If alleles for a certain locus are missing and there's only a delta
parameter, this means that all so far examined alleles behave properly and
can be offset safely. For those that also have mapping table in
\code{allele_from} and \code{allele_ref}, the mapping is done 1:1.
}
